backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Figma Integration
requests==2.31.0
httpx==0.25.2

# Website Capture & Analysis
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
cssselect==1.2.0

# Image Processing & Comparison
Pillow==10.1.0
opencv-python-headless==4.8.1.78
scikit-image==0.22.0
imagehash==4.3.1
numpy==1.26.2

# CSS Analysis
cssutils==2.9.0
tinycss2==1.2.1

# Color Analysis
extcolors==1.0.0
webcolors==1.13

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-multipart==0.0.6
python-json-logger==2.0.7
aiofiles==23.2.1

# Report Generation
jinja2==3.1.2
reportlab==4.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
